Updated Godot to the version 3.4.3-stable.

+++ godot/core/io/http_client.h
@@ -5,8 +5,8 @@
 /*                           GODOT ENGINE                                */
 /*                      https://godotengine.org                          */
 /*************************************************************************/
-/* Copyright (c) 2007-2021 Juan Linietsky, Ariel Manzur.                 */
-/* Copyright (c) 2014-2021 Godot Engine contributors (cf. AUTHORS.md).   */
+/* Copyright (c) 2007-2022 Juan Linietsky, Ariel Manzur.                 */
+/* Copyright (c) 2014-2022 Godot Engine contributors (cf. AUTHORS.md).   */
 /*                                                                       */
 /* Permission is hereby granted, free of charge, to any person obtaining */
 /* a copy of this software and associated documentation files (the       */
@@ -174,8 +174,8 @@ private:
 	Vector<uint8_t> chunk;
 	int chunk_left;
 	bool chunk_trailer_part;
-	int body_size;
-	int body_left;
+	int64_t body_size;
+	int64_t body_left;
 	bool read_until_eof;
 
 	Ref<StreamPeerTCP> tcp_connection;
